Some examples moved
