+++ examples3/src/main/scala/sttp/tapir/examples3/HelloWorldNettyFutureServer.scala
@@ -1,9 +1,9 @@
-package sttp.tapir.examples
+package sttp.tapir.examples3
 
 import sttp.client3.{HttpURLConnectionBackend, Identity, SttpBackend, UriContext, asStringAlways, basicRequest}
 import sttp.model.StatusCode
 import sttp.tapir.server.netty.{NettyFutureServer, NettyFutureServerBinding}
-import sttp.tapir.{PublicEndpoint, endpoint, query, stringBody}
+import sttp.tapir.*
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration.Duration
